+++ PicoXLSX/Properties/AssemblyInfo.cs
@@ -7,7 +7,7 @@ using System.Runtime.InteropServices;
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
 [assembly: AssemblyTitle ("PicoXLSX")]
-[assembly: AssemblyDescription ("PicoXLSX is a library to generate XLSX files in an easy and native way")]
+[assembly: AssemblyDescription("PicoXLSX is a library to generate Microsoft Excel files (XLSX) in an easy and native way")]
 [assembly: AssemblyConfiguration ("")]
 [assembly: AssemblyCompany("Raphael Stoeckli")]
 [assembly: AssemblyProduct ("PicoXLSX")]
@@ -30,5 +30,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2.1.1")]
-[assembly: AssemblyFileVersion("2.1.1")]
+[assembly: AssemblyVersion("2.2.0")]
+[assembly: AssemblyFileVersion("2.2.0")]